+++ lib/cunning_document_scanner.dart
@@ -1,5 +1,6 @@
 // Cunning Document Scanner - A simple document scanning library for Flutter.
 
+export 'src/android_scanner_mode.dart';
 export 'src/cunning_document_scanner.dart';
 export 'src/exceptions.dart';
 export 'src/ios_image_format.dart';